chroot: jenkins is tripping over these echo processs...

+++ scripts/chroot.sh
@@ -1148,12 +1148,16 @@ if [ "x${chroot_directory}" = "xenable" ]; then
 	du -h --max-depth=0 ${DIR}/deploy/${export_filename}/${deb_arch}-rootfs-${deb_distribution}-${deb_codename}
 else
 	cd ${tempdir}
-	packaging_keep_alive & PKG_KEEP_ALIVE_PID=$!
+	if [ ! -f ${DIR}/jenkins.build ] ; then
+		packaging_keep_alive & PKG_KEEP_ALIVE_PID=$!
+	fi
 	echo "Log: packaging rootfs: [${deb_arch}-rootfs-${deb_distribution}-${deb_codename}.tar]"
 	sudo LANG=C tar --numeric-owner -cf ${DIR}/deploy/${export_filename}/${deb_arch}-rootfs-${deb_distribution}-${deb_codename}.tar .
 	cd ${DIR}/
 	ls -lh ${DIR}/deploy/${export_filename}/${deb_arch}-rootfs-${deb_distribution}-${deb_codename}.tar
-	[ -e /proc/$PKG_KEEP_ALIVE_PID ] && sudo kill $PKG_KEEP_ALIVE_PID
+	if [ ! -f ${DIR}/jenkins.build ] ; then
+		[ -e /proc/$PKG_KEEP_ALIVE_PID ] && sudo kill $PKG_KEEP_ALIVE_PID
+	fi
 fi
 
 sudo chown -R ${USER}:${USER} ${DIR}/deploy/${export_filename}/
@@ -1174,9 +1178,13 @@ keep_alive () {
 if [ "x${chroot_tarball}" = "xenable" ] ; then
 	echo "Creating: ${export_filename}.tar"
 	cd ${DIR}/deploy/
-	keep_alive & KEEP_ALIVE_PID=$!
+	if [ ! -f ${DIR}/jenkins.build ] ; then
+		keep_alive & KEEP_ALIVE_PID=$!
+	fi
 	tar cvf ${export_filename}.tar ./${export_filename}
-	[ -e /proc/$KEEP_ALIVE_PID ] && sudo kill $KEEP_ALIVE_PID
+	if [ ! -f ${DIR}/jenkins.build ] ; then
+		[ -e /proc/$KEEP_ALIVE_PID ] && sudo kill $KEEP_ALIVE_PID
+	fi
 	cd ${DIR}/
 fi
 #
